+++ stats.h
@@ -1,4 +1,4 @@
 int Get_Uptime(unsigned long *Seconds);
 int Get_Load(double *Load);
-int Get_Memory(unsigned long *Memory);
+int Get_Memory(long double *Memory);
 int Get_Net(unsigned long *Sent, unsigned long *Recv, unsigned long *TotalSent, unsigned long *TotalRecv);
